clean up

+++ controllers/shopsController.js
@@ -16,7 +16,6 @@ const getPublicShopPage = asyncHandler(async(req,res)=>{
   if(!id) return res.status(400).json({message:'Shop id required'});
   const shop = await Shop.findById(id).select('-user -appointments').lean().exec();
   if(!shop) return res.status(400).json({message:`no shop found under id ${id}`});
-  console.log(shop);
   res.json({...shop});
 });
 
@@ -28,10 +27,8 @@ const makeAppointment = asyncHandler(async(req,res)=>{
     const {id} = req.params;
     const {name,lastName,service,date} = req.body;
     if(!id || !name || !lastName || !service || !date) return res.status(400).json({message:' All fields required'});
-    console.log('passed verification')
     const shop = await Shop.findById(id);
     if(!shop) return res.status(400).json({message:`no shop under id ${id}`});
-    console.log("shop found")
     //TODO check if customers appointment allready exists
     const customerInfo = {
       shopId :id,
@@ -40,9 +37,7 @@ const makeAppointment = asyncHandler(async(req,res)=>{
       date,
       active:true
     }
-    console.log("Appinfo created");
     const appointment = await Appointment.create({...customerInfo});
-    console.log('app created')
    
     if(!appointment) res.status(400).json({message:'Invalid appointment data'});
     shop.appointments.push(appointment);
@@ -53,26 +48,40 @@ const makeAppointment = asyncHandler(async(req,res)=>{
 
 })
 
-//@desc Get all Shops
+//@desc Get all Shops for a specific user
 //@route GET /shops
 //@access Private
 
-
-//TODO implement roles check ! if admin return all shops if shopkeeper check his id and return HIS shops
-const getAllShops = asyncHandler(async (req, res) => {
-  const shops = await Shop.find().lean().populate('user');
+const getMyShops = asyncHandler(async (req, res) => {
+  const id = req._id;
+  if(!id) return res.status(401).json({message:'Unauthorized. Login required'});
+  const user  = await User.findById(id);
+  if(!user) return res.status(400).json({message:'No such User'});
+  //if the user is admin return all shops
+  if(user.roles.includes('1000'))
+  {
+    const shops = await Shop.find().lean().exec();
+    if(!shops) return res.status(400).json({message:'No shops in db'});
+    const shopsWithUser = await Promise.all(shops.map(async (shop)=>{
+      const user = await User.findById(shop.user).lean().exec();
+      return {...shop,shopkeeper:user.username}
+    }));
+    return res.json(shopsWithUser);
+  }
+  //if the user is not an admin return his shops if any
+  const shops = await Shop.find({user:id}).lean().exec();
   //if no shops are found
   if (!shops?.length) {
     return res.status(400).json({ message: "No shops found" });
   }
 
   //populate shopkeeper in each shop before sending the response
-  // const shopsWithUser = await Promise.all(shops.map(async (shop)=>{
-  //   const user = await User.findById(shop.user).lean().exec();
-  //   return {...shop,username:user.username};
-  //  }))
+   const shopsWithUser = await Promise.all(shops.map(async (shop)=>{
+     const user = await User.findById(shop.user).lean().exec();
+     return {...shop,shopkeeper:user.username};
+    }))
 
-res.json(shops);
+res.json(shopsWithUser);
 });
 
 //@desc create new Shop
@@ -80,14 +89,15 @@ res.json(shops);
 //@access Private
 
 const createShop = asyncHandler(async (req, res) => {
-  const { user, title, description } = req.body;
-  console.log(user,title,description)
+  const id = req._id;
+  if(!id) return res.status(401).json({message:'Unauthorized. Log in required'});
+  const {title, description } = req.body;
   //verify data
-  if (!user || !title || !description) {
+  if (!title || !description) {
     return res.status(400).json({ message: "All input fields are required" });
   }
   //check if user exists in db LATER : if user has roles shopkeeper
-  const shopkeeper = await User.findById(user).lean().exec();
+  const shopkeeper = await User.findById(id).lean().exec();
   if (!shopkeeper) {
     return res
       .status(400)
@@ -109,11 +119,12 @@ const createShop = asyncHandler(async (req, res) => {
       .json({ message: `Shop with title ${title} allready exists` });
   }
   //if not duplicate create and store the shop
-  const shop = await Shop.create({ user, title, description });
+  const shop = await Shop.create({ user:id, title, description });
   //check if created successfully
   if (shop) {
     const publicLink = `http://localhost:3000/shops/public/${shop._id}`
-    await shop.publicLink.push(publicLink);
+    shop.publicLink = publicLink;
+    await shop.save();
     return res.status(201).json({ message: `Shop with title ${title} was created successfully` });
   } else {
     return res.status(400).json({ message: "Invalid Shop info" });
@@ -125,6 +136,8 @@ const createShop = asyncHandler(async (req, res) => {
 //@access Private
 
 const updateShop = asyncHandler(async (req, res) => {
+  const userId = req._id;
+  if(!userId) return res.status(401).json({message:'Unauthorized. Login required'});
   const { id, title, description } = req.body;
   //verify data
   if (!id || !title || !description) {
@@ -135,6 +148,7 @@ const updateShop = asyncHandler(async (req, res) => {
   if (!shop) {
     return res.status(400).json({ message: `No shop under id ${id} exists` });
   }
+  if(userId !==shop.user.toString()) return res.status(401).json({message:'unauthorized on this action'});
   //chech for duplicate title
   const duplicate = await Shop.findOne({ title }).lean().exec();
   if (duplicate && duplicate._id !== id) {
@@ -154,6 +168,9 @@ const updateShop = asyncHandler(async (req, res) => {
 //@access Private
 
 const deleteShop = asyncHandler(async (req, res) => {
+  const userId = req._id;
+ 
+  if(!userId) return res.status(401).json({message:'Unauthorized. Login required'})
     const {id} = req.body;
     //verify data
     if(!id){
@@ -165,13 +182,15 @@ const deleteShop = asyncHandler(async (req, res) => {
     {
         return res.status(400).json({message:`Shop with ${id} was not found`});
     }
+   
+    if(shop.user.toString() !==userId) return res.status(401).json({message:'Unauthorized on this action'})
     const result  = await shop.deleteOne();
     res.json({message:`Shop ${result.title} was delete successfully`});
 
 });
 
 module.exports = {
-  getAllShops,
+  getMyShops,
   createShop,
   updateShop,
   deleteShop,
